+++ core/network/src/main/java/com/jparkbro/network/ranking/RankingApi.kt
@@ -11,7 +11,7 @@ interface RankingApi {
     @GET(ApiConstants.RANK_REAL_TIME)
     suspend fun getRealTimeRanking(
         @Query("genre") genre: String?,
-        @Query("lastId") lastId: Int?,
+        @Query("lastId") lastId: Long?,
         @Query("lastValue") lastValue: Int?,
         @Query("size") size: Int?,
     ): Response<ApiResponse<RankingResponse>>
@@ -21,7 +21,7 @@ interface RankingApi {
         @Query("year") year: String?,
         @Query("season") season: Int?,
         @Query("genre") genre: String?,
-        @Query("lastId") lastId: Int?,
+        @Query("lastId") lastId: Long?,
         @Query("lastRank") lastRank: Int?,
         @Query("size") size: Int?,
     ): Response<ApiResponse<RankingResponse>>
@@ -29,7 +29,7 @@ interface RankingApi {
     @GET(ApiConstants.RANK_ALL_TIME)
     suspend fun getAllTimeRanking(
         @Query("genre") genre: String?,
-        @Query("lastId") lastId: Int?,
+        @Query("lastId") lastId: Long?,
         @Query("lastRank") lastRank: Int?,
         @Query("size") size: Int?,
     ): Response<ApiResponse<RankingResponse>>
